+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - Nothing yet
 
+## 5.1.2 (2024-08-01)
+
+- update dependencies following updated dependencies on udata [#470](https://github.com/datagouv/udata-front/pull/470)
+- update `@datagouv/components` and `vue-i18n` and fix report translation [#471](https://github.com/datagouv/udata-front/pull/471)
+
 ## 5.1.0 (2024-07-30)
 
 - Add organization edit to new admin [#412](https://github.com/datagouv/udata-front/pull/412)